Fix macOS compiler warning in HdrHistogram::makeIterator
Fix compiler warning:
"comparison of unsigned expression < 0 is always false"
Which was ocouring to to an incorrect cast from int64_t to uint64_t.

Reviewed-on: http://review.couchbase.org/119944
Tested-by: Build Bot <build@couchbase.com>

+++ utilities/hdrhistogram.cc
@@ -169,9 +169,8 @@ HdrHistogram::Iterator HdrHistogram::makeIterator(
     case Iterator::IterMode::Log:
         return makeLogIterator(1, 2);
     case Iterator::IterMode::Linear: {
-        uint64_t bucketWidth = static_cast<int64_t>(getMaxTrackableValue() -
-                                                    getMinTrackableValue()) /
-                               20;
+        int64_t bucketWidth =
+                (getMaxTrackableValue() - getMinTrackableValue()) / 20;
         if (bucketWidth < 0) {
             bucketWidth = 1;
         }
